src/telegram_bot/requirements.txt
fastapi
pydantic-settings
gunicorn
uvicorn
aiohttp
python-telegram-bot
asyncpg
